add inouts to DFS and BFS. Redisgn

+++ graph_lib/include/s21_graph_algorithms.h
@@ -15,20 +15,20 @@ namespace s21 {
 class GraphAlgorithms {
  public:
   struct TsmResult {
-    std::vector<int> vertices;
+    std::vector<float> vertices;
     double distance;
   };
   // Part 1
-  static std::list<float> DepthFirstSearch(Graph &graph, int start_vertex);
-  static std::list<float> BreadthFirstSearch(Graph &graph, int start_vertex);
+  static std::list<float> DepthFirstSearch(Graph &graph, float start_vertex);
+  static std::list<float> BreadthFirstSearch(Graph &graph, float start_vertex);
 
   // Part 2
-  static int GetShortestPathBetweenVertices(Graph &graph, int vertex1,
-                                            int vertex2);
-  static matrix<int> GetShortestPathsBetweenAllVertices(Graph &graph);
+  static float GetShortestPathBetweenVertices(Graph &graph, float vertex1,
+                                            float vertex2);
+  static matrix<float> GetShortestPathsBetweenAllVertices(Graph &graph);
 
   // Part 3
-  static matrix<int> GetLeastSpanningTree(Graph &graph);
+  static matrix<float> GetLeastSpanningTree(Graph &graph);
 
   // Part 4
   static TsmResult SolveTravelingSalesmanProblem(Graph &graph_test_);
@@ -46,7 +46,7 @@ class GraphAlgorithms {
   static bool Contains(const s21::stack<float> &stack, float elem);
   static bool Contains(const s21::queue<float> &queue, float elem);
   static bool Contains(const std::list<float> &list, float elem);
-  static void UpdateCost(int vertex, matrix<float> &adjacency_matrix,
+  static void UpdateCost(float vertex, matrix<float> &adjacency_matrix,
                          s21::queue<float> &adjacent_vertices,
                          std::list<float> &visited_vertices,
                          std::map<float, float> &path_cost);
