+++ graph_lib/source/graph_algorithms/bfs_dfc.cc
@@ -2,8 +2,8 @@
 
 namespace s21 {
 std::list<float> GraphAlgorithms::DepthFirstSearch(Graph& graph,
-                                                 int start_vertex) {
-  int vertex = start_vertex;
+                                                 float start_vertex) {
+  float vertex = start_vertex;
   matrix<float> adjacency_matrix = graph.GetGraph();
   stack<float> adjacent_vertices;
   std::list<float> visited_vertices{static_cast<float>(start_vertex)};
@@ -23,9 +23,9 @@ std::list<float> GraphAlgorithms::DepthFirstSearch(Graph& graph,
 
   return visited_vertices;
 }
-//
+
 std::list<float> GraphAlgorithms::BreadthFirstSearch(Graph& graph,
-                                                   int start_vertex) {
+                                                   float start_vertex) {
   matrix<float> adjacency_matrix = graph.GetGraph();
   s21::queue<float> adjacent_vertices;
   std::list<float> visited_vertices{start_vertex};
