+++ graph_lib/source/graph_algorithms/minimum_spanning_tree.cc
@@ -1,7 +1,7 @@
 #include "s21_graph_algorithms.h"
 
 namespace s21 {
-matrix<int> GraphAlgorithms::GetLeastSpanningTree(Graph& graph) {
+matrix<float> GraphAlgorithms::GetLeastSpanningTree(Graph& graph) {
   return {};
 }
 }  // namespace s21
